+++ pycf/cfl.pxd
@@ -194,11 +194,11 @@ cdef extern from "../../cfl/include/cfl_h_fit.h":
     void eshfit_data_free(eshfit_data *data)
     meshfit_data *meshfit_data_alloc(int n, eshfit_data **eshfit_d)
     void meshfit_data_free(meshfit_data *data)
-    double efit_obj(int n, double *x, double *grad, void *data) nogil
-    double mhfit_obj(int n, double *x, double *grad, void *data) nogil
-    double eshfit_obj(int n, double *x, double *grad, void *data) nogil 
-    double eshfit_hpro_obj(int n, double *x, double *grad, void *data) nogil 
-    double meshfit_obj(int n, double *x, double *grad, void *data) nogil
+    double efit_obj(size_t n, double *x, double *grad, void *data) nogil
+    double mhfit_obj(size_t n, double *x, double *grad, void *data) nogil
+    double eshfit_obj(size_t n, double *x, double *grad, void *data) nogil 
+    double eshfit_hpro_obj(size_t n, double *x, double *grad, void *data) nogil 
+    double meshfit_obj(size_t n, double *x, double *grad, void *data) nogil
     void efit_chi2(double *x, void *data, double *chi2) nogil
     void mhfit_chi2(double *x, void *data, double *chi2) nogil 
     void eshfit_chi2(double *x, void *data, double *chi2) nogil
